refactor(packages): collapse config/quality/orchestration into databox
Four workspace packages (databox-config, databox-quality,
databox-orchestration, databox-sources) become two: databox (shared
library: config, quality, orchestration) and databox-sources
(per-source dlt code + registry).

- git mv preserves history for settings, pipeline_config, engine,
  definitions, metrics
- root package renamed databox-workspace to free the databox name
- import paths rewritten (databox_config -> databox.config, etc.)
  across 11 .py files; Dagster resource string key "databox_config"
  preserved
- PROJECT_ROOT / TRANSFORMS_PATH parent counts bumped for deeper
  nesting
- README, CLAUDE.md, Taskfile, docs updated

All 30 tests pass, Dagster loads 24 assets, MkDocs builds clean,
SQLMesh plan clean.

Closes ticket:collapse-packages.
