--- packages/databox-config/pyproject.toml
@@ -1,14 +1,0 @@
-[project]
-name = "databox-config"
-version = "0.2.0"
-requires-python = ">=3.12"
-dependencies = [
-    "pydantic>=2.10.4",
-    "pydantic-settings>=2.13.0",
-    "python-dotenv>=1.0.1",
-    "pyyaml>=6.0",
-]
-
-[build-system]
-requires = ["setuptools>=70.0.0"]
-build-backend = "setuptools.build_meta"
