+++ packages/databox-sources/pyproject.toml
@@ -3,7 +3,7 @@ name = "databox-sources"
 version = "0.2.0"
 requires-python = ">=3.12"
 dependencies = [
-    "databox-config",
+    "databox",
     "dlt[duckdb]>=1.14.1",
     "pendulum>=3.0.0",
     "requests>=2.32.3",
@@ -15,4 +15,4 @@ requires = ["setuptools>=70.0.0"]
 build-backend = "setuptools.build_meta"
 
 [tool.uv.sources]
-databox-config = { workspace = true }
+databox = { workspace = true }
