+++ packages/databox/databox/config/pipeline_config.py
@@ -5,7 +5,7 @@ from typing import Any
 import yaml
 from pydantic import BaseModel, Field
 
-from databox_config.settings import PROJECT_ROOT
+from databox.config.settings import PROJECT_ROOT
 
 SOURCES_DIR = PROJECT_ROOT / "packages" / "databox-sources" / "databox_sources"
