+++ packages/databox/pyproject.toml
@@ -1,10 +1,15 @@
 [project]
-name = "databox-orchestration"
+name = "databox"
 version = "0.2.0"
+description = "Databox shared library — config, quality engine, and Dagster orchestration."
 requires-python = ">=3.12"
 dependencies = [
-    "databox-config",
     "databox-sources",
+    "pydantic>=2.10.4",
+    "pydantic-settings>=2.13.0",
+    "python-dotenv>=1.0.1",
+    "pyyaml>=6.0",
+    "soda-duckdb>=3.0.0",
     "dagster>=1.5.0",
     "dagster-webserver>=1.5.0",
     "dagster-dlt>=0.1.0",
@@ -18,6 +23,5 @@ requires = ["setuptools>=70.0.0"]
 build-backend = "setuptools.build_meta"
 
 [tool.uv.sources]
-databox-config = { workspace = true }
 databox-sources = { workspace = true }
 dagster-sqlmesh = { git = "https://github.com/Doctacon/dagster-sqlmesh.git" }
